Tests: Classic Editor: No Forms: Refactor
