--- tests/EndToEnd/forms/post-types/PageNoFormCest.php
@@ -1,72 +1,0 @@
-<?php
-
-namespace Tests\EndToEnd;
-
-use Tests\Support\EndToEndTester;
-
-/**
- * Tests for WordPress Pages when no Kit Forms exist in the Kit account.
- *
- * @since   1.9.6.1
- */
-class PageNoFormCest
-{
-	/**
-	 * Run common actions before running the test functions in this class.
-	 *
-	 * @since   1.9.6.1
-	 *
-	 * @param   EndToEndTester $I  Tester.
-	 */
-	public function _before(EndToEndTester $I)
-	{
-		// Activate and Setup Kit plugin.
-		$I->activateKitPlugin($I);
-		$I->setupKitPluginCredentialsNoData($I);
-		$I->setupKitPluginResourcesNoData($I);
-	}
-
-	/**
-	 * Test that UTM parameters are included in links displayed in the metabox for the user to sign in to
-	 * their Kit account.
-	 *
-	 * @since   1.9.6
-	 *
-	 * @param   EndToEndTester $I  Tester.
-	 */
-	public function testUTMParametersExist(EndToEndTester $I)
-	{
-		// Activate Classic Editor Plugin.
-		$I->activateThirdPartyPlugin($I, 'classic-editor');
-
-		// Navigate to Pages > Add New.
-		$I->amOnAdminPage('post-new.php?post_type=page');
-
-		// Check that no PHP warnings or notices were output.
-		$I->checkNoWarningsAndNoticesOnScreen($I);
-
-		// Check that the metabox is displayed.
-		$I->seeElementInDOM('#wp-convertkit-meta-box');
-
-		// Confirm that UTM parameters exist for the 'sign in to Kit' link.
-		$I->seeInSource('<a href="https://app.kit.com/?utm_source=wordpress&amp;utm_term=en_US&amp;utm_content=convertkit" target="_blank">sign in to Kit</a>');
-
-		// Deactivate Classic Editor Plugin.
-		$I->deactivateThirdPartyPlugin($I, 'classic-editor');
-	}
-
-	/**
-	 * Deactivate and reset Plugin(s) after each test, if the test passes.
-	 * We don't use _after, as this would provide a screenshot of the Plugin
-	 * deactivation and not the true test error.
-	 *
-	 * @since   1.9.6.7
-	 *
-	 * @param   EndToEndTester $I  Tester.
-	 */
-	public function _passed(EndToEndTester $I)
-	{
-		$I->deactivateKitPlugin($I);
-		$I->resetKitPlugin($I);
-	}
-}
